html/browsers/browsing-the-web/back-forward-cache/service-worker-unregister.https.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/utils.js"></script>
<script src="/common/dispatcher/dispatcher.js"></script>
<script src="resources/helper.sub.js"></script>
<script src="/service-workers/service-worker/resources/test-helpers.sub.js"></script>
<script>

promise_test(async t => {
  const workerUrl = 'resources/clients-matchall-service-worker.js?pipe=header(Service-Worker-Allowed,../)';
  const registration = await service_worker_unregister_and_register(t, workerUrl, './');
  t.add_cleanup(_ => registration.unregister());
  await wait_for_state(t, registration.installing, 'activated');

  const pageA = new RemoteContext(token());
  const pageB = new RemoteContext(token());

  const urlA = location.origin + executorPath + pageA.context_id;
  const urlB = originCrossSite + executorPath + pageB.context_id;

  window.open(urlA, '_blank', 'noopener');

  await pageA.execute_script(waitForPageShow);
  await pageA.execute_script(
    (url) => prepareNavigation(() => {
      location.href = url;
    }),
    [urlB]);

  await pageB.execute_script(waitForPageShow);

  await registration.unregister();

  await pageB.execute_script(
    () => {
      prepareNavigation(() => { history.back(); });
    }
  );

  await pageA.execute_script(waitForPageShow);
  await assert_bfcached(pageA);
}, 'Unregister service worker while a controlled page is in BFCache should not crash');
</script>
